# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'HaloSitek Core API'
  description: 'OpenAPI metadata for HaloSitek Backoffice API.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000/api/v1'
    description: 'Local Development Server'
paths:
...
  /faqs:
    get:
      tags:
        - FAQ
      summary: 'List FAQs'
      description: 'Returns FAQ items; public users can only see active entries.'
      operationId: 6349d195bc7a1708dc2ee3e3b72ca3c3
      responses:
        200:
          description: 'FAQ list retrieved successfully'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 200
                message: 'FAQ list retrieved successfully'
                data:
                  -
                    id: 01HZX9M1F45M2Z6K7T9K7Y8QFQ
                    question: 'How do I become an architect on this platform?'
                    answer: 'Register and complete your architect profile.'
                    is_active: true
                meta:
                  current_page: 1
                  last_page: 1
                  per_page: 15
                  total: 1
                links:
                  first_page_url: 'http://localhost:8000/api/v1/faqs?page=1'
                  last_page_url: 'http://localhost:8000/api/v1/faqs?page=1'
                  next_page_url: null
                  prev_page_url: null
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/ServerError'
      security: []
    post:
      tags:
        - FAQ
      summary: 'Create FAQ'
      description: 'Creates a new FAQ entry through admin access.'
      operationId: 7bbd0e966b08c1636271fb03d791977a
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - question
                - answer
              properties:
                question:
                  type: string
                answer:
                  type: string
                is_active:
                  type: boolean
              type: object
      responses:
        201:
          description: 'FAQ created successfully'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 201
                message: 'FAQ created successfully'
                data:
                  id: 01HZX9M1F45M2Z6K7T9K7Y8QFQ
                  question: 'How do I become an architect on this platform?'
                  answer: 'Register and complete your architect profile.'
                  is_active: true
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        422:
          $ref: '#/components/responses/ValidationError'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        -
          BearerAuth: []
  '/faqs/{id}':
    get:
      tags:
        - FAQ
      summary: 'Get FAQ detail'
      description: 'Returns details for a single FAQ by id.'
      operationId: c3783b4ea36d8234f5c7bfffae857f39
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'FAQ retrieved successfully'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 200
                message: 'FAQ retrieved successfully'
                data:
                  id: 01HZX9M1F45M2Z6K7T9K7Y8QFQ
                  question: 'How do I become an architect on this platform?'
                  answer: 'Register and complete your architect profile.'
                  is_active: true
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/ServerError'
      security: []
    put:
      tags:
        - FAQ
      summary: 'Update FAQ'
      description: 'Updates an existing FAQ entry by id.'
      operationId: 26f32aa1cf53af2bda80c4d3a0f4f760
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                question:
                  type: string
                answer:
                  type: string
                is_active:
                  type: boolean
              type: object
      responses:
        200:
          description: 'FAQ updated successfully'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 200
                message: 'FAQ updated successfully'
                data:
                  id: 01HZX9M1F45M2Z6K7T9K7Y8QFQ
                  question: 'How do I become an architect on this platform?'
                  answer: 'Your architect profile has been updated.'
                  is_active: true
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'
        422:
          $ref: '#/components/responses/ValidationError'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        -
          BearerAuth: []
    delete:
      tags:
        - FAQ
      summary: 'Delete FAQ'
      description: 'Deletes an FAQ entry by id.'
      operationId: 030910a5c473d7b6c66516490013e3b9
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'FAQ deleted successfully'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 200
                message: 'FAQ deleted successfully'
                data: null
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        -
          BearerAuth: []
...
components:
  schemas:
...
    ApiError:
      properties:
        success:
          type: boolean
          example: false
        status_code:
          type: integer
          example: 401
        message:
          type: string
          example: Unauthorized
        errors:
          type: object
          nullable: true
      type: object
  responses:
    UnauthorizedError:
      description: 'Unauthorized - missing or invalid access token.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: 'Unauthorized. Access token is missing or invalid.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
    ForbiddenError:
      description: 'Forbidden - user does not have permission for this action.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 403
                  message:
                    type: string
                    example: 'Forbidden. You are not allowed to perform this action.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
    NotFoundError:
      description: 'Not Found - requested resource was not found.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: 'Resource not found.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
...
    ValidationError:
      description: 'Validation Error - request input is invalid.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 422
                  message:
                    type: string
                    example: 'Validation error.'
                  errors:
                    type: object
                    example:
                      email:
                        - 'The email field is required.'
                      password:
                        - 'The password must be at least 8 characters.'
                type: object
    ServerError:
      description: 'Internal Server Error - an unexpected server error occurred.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: 'Internal server error.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
...
  securitySchemes:
    BearerAuth:
      type: http
      description: 'Use Sanctum access token in Authorization header as: Bearer {token}'
      bearerFormat: JWT
      scheme: bearer
...